requirements.txt
Flask
gunicorn
requests
graphlab-create